requirements.txt
requests
reportlab
google-auth
google-auth-oauthlib
google-api-python-client
PyYAML
google-generativeai
python-dotenv